requirements.api.txt
fastapi==0.110.0
uvicorn==0.27.1
tensorflow==2.20.0
keras==3.13.2
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.9